requirements.txt
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.5.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2
sqlalchemy>=2.0.25
PyJWT>=2.8.0
email-validator>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
python-multipart>=0.0.6
